Extract ColumnSelector
